move trade details
