+++ src/components/trade/swap/trade-details/trade-info.tsx
